+++ tests/testthat/test_survey_mean_factor.r
@@ -86,3 +86,26 @@ out_srvyr <- dstrata %>%
 test_that("survey_* preserves character when calculating a statistic (multi grps)",
           expect_true(class(out_srvyr$stype) == "character")
 )
+
+
+# confidence intervals
+out_survey_mn <- svymean(~awards, dstrata)
+out_survey_tot <- svytotal(~awards, dstrata)
+out_survey <- dplyr::data_frame(
+  awards = factor(c("No", "Yes")),
+  pct = as.numeric(out_survey_mn),
+  pct_low = as.numeric(confint(out_survey_mn)[, 1]),
+  pct_upp = as.numeric(confint(out_survey_mn)[, 2]),
+  tot = as.numeric(out_survey_tot),
+  tot_low = as.numeric(confint(out_survey_tot)[, 1]),
+  tot_upp = as.numeric(confint(out_survey_tot)[, 2])
+)
+
+out_srvyr <- dstrata %>%
+  group_by(awards) %>%
+  summarize(pct = survey_mean(vartype = "ci"),
+            tot = survey_total(vartype = "ci"))
+
+test_that(
+  "survey_mean and survey_total work with cis",
+  expect_equal(out_srvyr, out_survey))