+++ NEWS.md
@@ -2,3 +2,5 @@
 
 * New function `cascade` summarizes groups, and cascades to create
   summary statistics of groups of groups.
+
+* Fixed a bug for confidence intervals for `survey_total()` on groups.